14-github-agent/requirements.txt
python-dotenv
# PyGithub will be added here
# GitPython will be added here